Image Map 	Server - Grid.hbm.xml 		imagesOnGrid are no longer lazy loaded 	Client - ResizableWidget.java 		Changed the code that was setting style attributes to add a stylename and use CSS instead, it was causing exceptions 	Client - AddExistingImagePopup.java 		Now only displays images that have not already been added to the image map 	Client - AddImageDialog.java 		Fixed a typo, the Submit button was getting set to have the primary and secondary button CSS, now only gets primary and the cancel button gets secondary 	Client - OpacityPopup.java 		A line had 3 .getParent()'s in it, but should've been 2. This would cause an exception, removed one and opacity now works fine 	Client - OpacityListener.java 		Fixed a bug where a string was trying to be parsed into an int, it would fail because of decimals. Now parses to a double, then gets casted into an int 	Client ImageBrowserDetails.java 		Fixed a bug where the image would not change on zooming or panning, this was because the image needed to be added to the map<> of images on the grid 		Now saves the chemical analyses when clicking save. This will update their X,Y positions 	Client ImageBrowserMouseListener.java 		Now does not allow you to place a chemical analysis point outside its image 		slightly increased the pixel range you need to click near a chemical analysis point to drag it 	Client ImageList.java 		Now only displays images that are not already in use if you pass in a collection of images

+++ mpdb-client/src/edu/rpi/metpetdb/client/ui/image/browser/click/listeners/OpacityListener.java
@@ -30,10 +30,10 @@ public class OpacityListener implements ClickListener {
 			}
 			public void onSuccess(final Object result) {
 				if (result instanceof String) {
-					setOpacity(iog.getActualImage().getElement(), Integer
-							.parseInt((String) result));
-					setOpacity(iog.getImageContainer().getElement(), Integer
-							.parseInt((String) result));
+					setOpacity(iog.getActualImage().getElement(), (int) Double
+							.parseDouble((String) result));
+					setOpacity(iog.getImageContainer().getElement(),
+							(int) Double.parseDouble((String) result));
 				}
 			}
 		}.begin();
